Port 1abc689a32fc20b38227060e4da5e0934df8fc19 and e5d14162ff278c41e09c42769cc31401feeeba18 to art.
(The last pieces of 64-bit nio support?)

+++ src/jni_internal.cc
@@ -2143,7 +2143,7 @@ class JNI {
     // At the moment, the Java side is limited to 32 bits.
     CHECK_LE(reinterpret_cast<uintptr_t>(address), 0xffffffff);
     CHECK_LE(capacity, 0xffffffff);
-    jint address_arg = reinterpret_cast<jint>(address);
+    jlong address_arg = reinterpret_cast<jlong>(address);
     jint capacity_arg = static_cast<jint>(capacity);
 
     jobject result = env->NewObject(WellKnownClasses::java_nio_DirectByteBuffer,
